separate taskserver away from cmd/main adjust number of workers to 6 and Tasks channel capacity to 1 million
